Renaming the file with events as well
